Move ::constructor tests and new fromOptions tests to unit testing directory
